1. Create interrupt handling for ADCs 2. Create Event Handler 3. Create TIM5 for ADC input signal meas (TIM2 still for internal temp sensor)

+++ DigitalOscilloscope/include/Adc.h
@@ -12,7 +12,7 @@
 
 
 void Adc_Init();
-double Adc_Read();
+uint16_t Adc_Read();
 double Adc_ReadIntTemp();
 
 #endif /* ADC_H_ */
